fix upload avatar new user

+++ controllers/fileController.js
@@ -152,7 +152,7 @@ class FileController {
         try {
             const file = req.files.file
             const user = await User.findById(req.user.id)
-            if (user.avatar !== null) {
+            if (user.avatar !== null && user.avatar !== undefined) {
                 fs.unlinkSync(`${req.staticPath}/${user.avatar}`)
             }
             const avatarName = Uuid.v4() + '.jpg'
